migration: add special transcode task
* The TranscodeVideoTaskQuiet task was not in the list of tasks inside
  celery causing an error when trying to schedule. (closes #1337)
