+++ cds/modules/migrator/tasks.py
@@ -1,0 +1,46 @@
+# -*- coding: utf-8 -*-
+
+#
+# This file is part of CERN Document Server.
+# Copyright (C) 2017 CERN.
+#
+# CERN Document Server is free software; you can redistribute it
+# and/or modify it under the terms of the GNU General Public License as
+# published by the Free Software Foundation; either version 2 of the
+# License, or (at your option) any later version.
+#
+# CERN Document Server is distributed in the hope that it will be
+# useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with CERN Document Server; if not, write to the
+# Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston,
+# MA 02111-1307, USA.
+#
+# In applying this license, CERN does not
+# waive the privileges and immunities granted to it by virtue of its status
+# as an Intergovernmental Organization or submit itself to any jurisdiction.
+"""Record migration special."""
+
+from ..deposit.api import deposit_video_resolver
+from ..records.resolver import record_resolver
+from ..webhooks.tasks import TranscodeVideoTask
+
+
+class TranscodeVideoTaskQuiet(TranscodeVideoTask):
+    """Transcode without index or send sse messages."""
+
+    def on_success(self, *args, **kwargs):
+        # get deposit and record
+        deposit_id = args[3]['deposit_id']
+        video = deposit_video_resolver(deposit_id)
+        rec_video = record_resolver.resolve(video['recid'])[1]
+        # sync deposit --> record
+        video._sync_record_files(record=rec_video)
+        video.commit()
+        rec_video.commit()
+
+    def _update_record(self, *args, **kwargs):
+        pass
